parcelpy/streamlit/requirements.txt
# Streamlit and web interface
streamlit>=1.28.0
streamlit-folium>=0.15.0
streamlit-aggrid>=0.3.4
plotly>=5.15.0

# Core dependencies from database module
duckdb>=0.9.0
pandas>=1.5.0
geopandas>=0.12.0
shapely>=2.0.0
pyproj>=3.0.0
pyarrow>=8.0.0
fastparquet>=0.8.0

# Visualization dependencies from viz module
numpy>=1.20.0
matplotlib>=3.4.0
rasterio>=1.2.0
rioxarray>=0.10.0
xarray>=0.19.0
seaborn>=0.11.0
folium>=0.14.0
rasterstats>=0.16.0
affine>=2.3.0

# Additional utilities
tqdm>=4.64.0
psutil>=5.9.0
pyyaml>=6.0
simplejson>=3.17.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0 